Update collumn type of 'ip' in the curricula table migration

+++ database/migrations/2022_08_26_204852_create_curricula_table.php
@@ -15,7 +15,7 @@ class CreateCurriculaTable extends Migration
     {
         Schema::create('curricula', function (Blueprint $table) {
             $table->id();
-            $table->integer('ip');
+            $table->bigInteger('ip');
             $table->string('nome');
             $table->string('email');
             $table->bigInteger('telefone');
